pyproject.toml
[project]
name = "sdd-ollama"
version = "0.1.0"
requires-python = ">=3.12,<3.14"
dependencies = [
    "chromadb>=1.1.1",
    "crawl4ai>=0.8.6",
    "crewai-tools>=1.14.0",
    "curl-cffi>=0.15.0",
    "ddgs>=9.13.0",
    "jsonschema>=4.0",
    "langchain>=1.2.15",
    "langgraph>=1.1.9",
    "langsmith>=0.7.33",
    "playwright>=1.58.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "trafilatura>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: Testes que chamam I/O ou LLM real (integração)",
    "deterministic: Testes de lógica pura sem dependências externas",
    "spec_driven: Testes que validam contra a spec",
    "e2e: Testes fim-a-fim do pipeline",
    "integration: Testes de integração entre componentes",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["F401", "F841", "C901"]

[tool.ruff.lint.mccabe]
max-complexity = 12
